Use generic MerkleTreeImpl
